tvOS app icon fix, also icon revisions
Fixed the @2x icon sizing. Each @2x icon was previously the same size as its @1x counterpart, resulting in odd graphical errors when the icon was selected on the tvOS home screen.

Switched to the non-outlined RetroArch logo to match the iOS app icon—and because I was unable to find a high enough resolution version of the outlined logo to properly create @2x icon layers for it.

Separated the icon layers further, going from two layers to three (eyes on their own layer now) to improve the resultant tvOS home screen parallax effect on it. The eyes now move ever so slightly—independent of the body.