Fix possible never-ending while loops
'while(size)' will only ever be false and exit the loop if size = 0.  size is a variable of type 'snd_pcm_sframes_t' which can be positive or negative (long integer).

size is only decremented in the loops so if it goes below zero, the evaluation will always be true and we will become trapped in the loop until the process is killed!  Tested this on an ARM9 platform by changing both loops to while(1), and confirmed this will freeze RetroArch, allowing no input resulting in the application process having to be killed to resume.