(WiiU) aroma: fix conflict with FTPiiU plug-in
== DETAILS
The HID polling thread was written to be run on a single core, and
basically completely occupies it, not allowing any other threads to
run on it.

It turns out, this is a problem in Aroma, because other things can
be running on your cores. For instance, the FTPiiU server thread runs
on the same core as the HID polling thread, which explains why FTPiiU
stopped working when RetroArch was running.

The other problem, is that if FTPiiU is holding the IPC BufPool mutex
when RA is started, a deadlock will eventually happen because RA will
try to acquire the mutex and the FTPii thread, being suspended, won't
be able to release it.

So!

To fix it:
- add a 500-microsecond sleep to the HID polling thread
- update to WUT's HID headers, which makes some minor changes to
  structure member names