Fix runtime display. (#15581)
Due to an error in 631301b3f7d58d0969c1ebdaf2bb3907be5d0025, the
runtime display string was working well only when the label had
a length of 7. Even the English "Runtime:" was a bit truncated
as a space was supposed to be added after the :, but it was
even more noticeable in other languages.