(Language) Fix language detection in Canada to English
Canada's default language is often picked up as `en_CA.UTF-8`, which inadvertently gets discovered as `ca`, or Catalan. This means that if you're in Canada, you would end up getting Catalan as your default language.

Adding `en_CA` to the language list as British English fixes the language detection for Canadians. This must be indexed above Catalan's `ca` so that it uses English instead of Catalan.