Remove json_parser and use rc_json_parse_response
Some files of the webhooks directory have been updated or deleted. The json_parser.c and json_parser.h files were created for the RetroArch frontend, but they are more complex than the application needs. The application was already using RetroAchievements's rc_json_parse_response function in other parts of the code, so I've switched the json_parser to use the same function everywhere for consistency and simplicity.

Also, small fixes were made in webhooks_progress_downloader.c and webhooks_client.c to correctly pass variable values as strings, not as integers.

The change will make it easier to maintain the code and reduce complexity.