Remove Raspberry Pi userland
Userland was added at: https://github.com/libretro/RetroArch/pull/8132
However, its only use was with CRTSwitchRes together with VideoCore.
Over the times, the legacy VideoCore drivers have lost usage as
nowadays Mesa supports the VideoCore IV chip in RPi 0..3, and
RPi 4 and/or 64-bit was never supported.

Option is kept to use the legacy drivers, but it is now utilized
from the OS.