Refactor code in qb.moc.sh
A conditional statement within the script qb.moc.sh has been removed. Previously, this condition tested the successful creation of an executable by using the C++ compiler, and marked a flag "moc_works" as true if successful. The refactoring removed this condition, and the code now directly attempts to perform that action without checking for success.