Refactor moc.sh script in qb module
The INCLUDES printf call for QT5CORE_LIBS has been removed from the qb.moc.sh script. Additionally, the script now contains an explicit if-then condition to check the success of the printf call for 'CXX -o TEMP_EXE' and further commands. Successful completion of this command will now trigger the flagging of 'moc_works' as 1, eliciting an immediate break for efficiency.