Add FlatUX Theme, Remove Retired Themes (#16441)
* Add FlatUX Theme, Remove Retired Themes

With some help from @gouchi, I decided to try and add the FlatUX theme, which has been in the retroarch-assets repository for years.  FlatUX was designed to merge the FlatUI and Retroactive themes into a single, unified design.

With the addition of FlatUX, the Retroactive and Neoactive themes will be retired. This PR removes them both from the menus. The theme files will be removed from retroarch-assets, but are archived on baxy-retroarch-themes.

* Edited Language Files to Add FlatUX, Remove Retired Themes

Edited the language files to Add FlatUX and remove the Retroactive and Neoactive themes.

* Fixed Language Files Again

I apologize for my errors.  Let's see if this gets it right.

* Updated IT Language File