Error code framework
As one of the feedbacks from the RetroArch UI Improvements
series put together by Nick Watt, the ways RetroArch (or any
other frontend) could indicate errors about the cores, is quite
limited. The Libretro API functions provide little feedback
opportunity for the cores to indicate problems.

By introducing a new environment call, the possibility will be
available across the board. This change is a wireframe and may be
extended / adopted as needed.