File extension list generation rewrite
Only limit extensions under the condition that the valid_extensions list
is not empty and the user wishes to limit the extensions. Otherwise
present files of all recognized extensions.

Part of this logic was implemented in 8f3280129282706ce86a98d6 where the
point was to show supported files when core info file is missing. The
other part is the filter_by_current_core option.